Generate parse.js on prepublish
So no one will have to generate it themselves.

We need an empty `.npmignore` so npm uses that instead of the `gitignore` and pushes the generated `parse.js` when we publish.

This also moves Webpack to be a dev dep which might break Travis on Node 0.8 but we should really start to encourage our users to move away from that.