Bump nock to a version that is compatible with Node 10
I'm not entirely sure why, but on Node 10, with nock < 9.6.1,
if delayConnection() is called, it does not appear to be respected
by the timeout set when making a request throught the request
library.

Judging by the diff https://github.com/nock/nock/compare/v9.6.0...v9.6.1,
it's not clear why this is. But hey, it fixes the failing tests on
Node 10.