Support fetching channel info on trigger (#135)
* Support fetching channel info on trigger

Adds support for a new experimental feature of the Channels API:
Specifically the ability to specify an `info` parameter in a
trigger body, or with an event in a batch trigger body. The server
will respond with attributes in the same format as the `/channels`
endpoint.

The `info` param is specified in a "params" object, which is
passed in as the fourth parameter to the `trigger` method. The
`socket_id` parameter is specified in the same way.

Note: this introduces a breaking change for users that were
passing in a `socket_id`.

It is not a breaking change for `triggerBatch`.

`triggerBatch` has also been updated to support `socket_id` being
included as part of event objects.